# hazelnet-community/src/main/resources/hazelnet-community-openapi.yml
openapi: 3.0.0
info:
  version: '0.0.1'
  title: 'HazelNet Community'
servers:
  - url: 'localhost:8085'
paths:
...
  /snapshots/stake:
    post:
      summary: 'Schedule a token snapshot'
      operationId: scheduleSnapshot
      tags:
        - snapshot
        - token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiAssetSnapshot'
      responses:
        '201':
          description: 'Snapshot details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiAssetSnapshot'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /products:
    post:
      summary: Add Product
      operationId: addProduct
      tags:
        - claims
        - products
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhysicalProduct'
      responses:
        '201':
          description: 'Physical product'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalProduct'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /products/{productId}:
    get:
      summary: Get Product Info
      operationId: getProduct
      tags:
        - claims
        - products
      parameters:
        - $ref: '#/components/parameters/ProductIdPathParameter'
      responses:
        '200':
          description: 'Physical Product'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalProduct'
...
components:
  parameters:
...
    ProductIdPathParameter:
      name: productId
      in: path
      description: The ID of this product
      required: true
      style: simple
      example: 62362
      schema:
        type: integer
        format: int64
...
  schemas:
...
    PhysicalProduct:
      type: object
      required:
        - name
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        name:
          type: string
          example: T-Shirt
        variations:
          type: object
          example:
            colors:
              - Green
              - Black
              - Red
            sizes:
              - S
              - M
              - L
              - XL
              - XXL
...
    MultiAssetSnapshot:
      type: object
      required:
        - snapshotTime
        - policyId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        createTime:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T13:20:57Z
        snapshotTime:
          type: string
          format: date-time
          example: 2022-11-10T13:20:57Z
        policyId:
          $ref: '#/components/schemas/TokenPolicyId'
        assetFingerprint:
          $ref: '#/components/schemas/AssetFingerprint'
        taken:
          readOnly: true
          type: boolean
...
    TokenPolicyId:
      type: string
      description: The Policy ID
      example: ceb5dedd6cda3f0b4a98919b5d3827e15e324771642b57e0e6aabd57
      pattern: '^[A-Za-z0-9]{56}$'
    AssetFingerprint:
      type: string
      description: The fingerprint of an asset
      example: asset1ls9tt8s03x2gcqjyppx346ct64jts7fnwsmequ
      pattern: '^asset1[A-Za-z-0-9]{38}$'
...
    CreatedResponseHeader:
      type: string
      format: uri
      description: URL of the newly created resource
...